Split Reader on two classes
