+++ Kitodo/src/main/java/org/kitodo/forms/TemplateForm.java
@@ -39,7 +39,7 @@ import org.kitodo.helper.Helper;
 import org.kitodo.helper.SelectItemList;
 import org.kitodo.model.LazyDTOModel;
 import org.kitodo.services.data.TaskService;
-import org.kitodo.workflow.model.Reader;
+import org.kitodo.workflow.model.Converter;
 
 @Named("TemplateForm")
 @SessionScoped
@@ -152,8 +152,8 @@ public class TemplateForm extends TemplateBaseForm {
         if (Objects.nonNull(this.template.getTitle()) && !this.template.getTitle().isEmpty()) {
             try {
                 if (this.template.getTasks().isEmpty()) {
-                    Reader reader = new Reader(this.template.getWorkflow().getFileName());
-                    this.template = reader.convertWorkflowToTemplate(this.template);
+                    Converter converter = new Converter(this.template.getWorkflow().getFileName());
+                    converter.convertWorkflowToTemplate(this.template);
                 }
             } catch (DAOException e) {
                 Helper.setErrorMessage("errorDiagramConvert", new Object[] {this.template.getWorkflow().getTitle() }, logger, e);
